Fixed region serialization, improved region coloring
