+++ lib/action_region_editor_component.dart
@@ -1,4 +1,5 @@
 import 'package:angular2/core.dart';
+import 'package:code_steps/action_region.dart';
 import 'package:code_steps/lesson_editor_component.dart';
 import 'package:code_steps/lesson_serializer.dart';
 import 'package:code_steps/step_action.dart';
@@ -28,13 +29,6 @@ class ActionRegionEditorComponent implements OnChanges {
   StepActionsProvider stepActionsProvider;
   StepContextService stepContextService;
   EnumStringHelper<StepActionType> esh;
-  Set<StepActionType> get currentStepActions {
-    if (!region.stepData.containsKey(stepContextService.stepIndex)) {
-      region.stepData[stepContextService.stepIndex] =
-          new Set<StepActionType>();
-    }
-    return region.stepData[stepContextService.stepIndex];
-  }
 
   ActionRegionEditorComponent(
       this.stepActionsProvider, this.stepContextService) {
@@ -43,31 +37,23 @@ class ActionRegionEditorComponent implements OnChanges {
     esh = new EnumStringHelper<StepActionType>();
   }
 
+  Set<StepActionType> get currentStepActions => region.stepData[stepContextService.stepIndex];
+  set currentStepActions(Set<StepActionType> actions) => region.stepData[stepContextService.stepIndex] = actions;
+
   void deleteRegion() {
     onDelete.emit(region.marker.id);
     region = null;
   }
 
-  Map<StepActionType, bool> _actionUIToggles;
-  Map<StepActionType, bool> get actionUIToggles {
-    _actionUIToggles ??= {};
-    currentStepActions.forEach((t) => _actionUIToggles[t] = true);
-    StepActionType.values
-        .where((t) => !_actionUIToggles.containsKey(t))
-        .forEach((t) => _actionUIToggles[t] = false);
-    return _actionUIToggles;
-  }
+  Map<StepActionType, bool> get actionUIToggles => region.getActionStates(stepContextService.stepIndex);
 
   void updateSelected(StepActionType type, bool event) {
-    onDataChange.emit(actionUIToggles);
-    actionUIToggles[type] = event;
+    if (currentStepActions == null) currentStepActions = new Set<StepActionType>();
     event ? currentStepActions.add(type) : currentStepActions.remove(type);
+    onDataChange.emit(actionUIToggles);
   }
 
   @override // called when @input() region is changed
   void ngOnChanges(Map<String, SimpleChange> changes) {
-    if(changes.containsKey('region')) {
-      _actionUIToggles = {};
-    }
   }
 }
