+++ lib/lesson_editor_component.dart
@@ -5,6 +5,7 @@ import 'dart:js';
 import 'package:angular2/core.dart';
 import 'package:ace/ace.dart' as ace;
 import 'package:ace/proxy.dart';
+import 'package:code_steps/action_region.dart';
 import 'package:code_steps/action_region_editor_component.dart';
 import 'package:code_steps/lesson_serializer.dart';
 import 'package:code_steps/step_action.dart';
@@ -116,26 +117,6 @@ class AceEditorComponent implements OnInit {
   set dom_id(id) => elementRef.nativeElement.id = id;
 }
 
-class ActionRegion {
-  ace.Marker marker;
-  String uniqClass;
-  Map<int, Set<StepActionType>> stepData = {};
-
-  ActionRegion(this.marker, this.uniqClass);
-
-  String toString() => "ActionEditRegion($marker, $stepData)";
-
-  Map toJson() {
-    var obj = {
-      'range': marker.range,
-      'step_data': LessonSerializer.transformMap(stepData,
-          valuesTransformer: (Set<StepActionType> set) =>
-              set.map(LessonSerializer.stepActionTypeTransformer).toList())
-    };
-    print(obj);
-    return obj;
-  }
-}
 
 @Component(selector: 'ace-code-edit', template: '', styles: const [
   '''
@@ -168,8 +149,8 @@ class LessonCodeEditorComponent extends AceEditorComponent implements OnInit {
     regions.forEach((region_map) {
       ActionRegion r = addActionMarker(region_map['range'])
         ..stepData = LessonSerializer.transformMap(region_map['step_data'],
-            keysTransformer: LessonSerializer.destringifyInt,
-            valuesTransformer: (v) => v.toSet());
+            key: LessonSerializer.destringifyInt,
+            value: (v) => v.toSet());
       print('=====serialized region output=====');
       print(r);
     });
@@ -222,7 +203,8 @@ class LessonCodeEditorComponent extends AceEditorComponent implements OnInit {
   static const red = const Color(126, 13, 13, 0.68);
   static const purple = const Color(197, 23, 158, 0.25);
   static const yellow = const Color(79, 76, 15, 0.66);
-  recolorActiveRegion(Map<StepActionType, bool> typeEnabledState) {
+  recolorRegion(
+      ActionRegion region, Map<StepActionType, bool> typeEnabledState) {
     Color c = null; // todo experimental
     Function addColor =
         (Color base, Color toAdd) => base == null ? toAdd : toAdd + base;
@@ -233,16 +215,14 @@ class LessonCodeEditorComponent extends AceEditorComponent implements OnInit {
       c = addColor(c, red);
     }
     if (typeEnabledState[StepActionType.Spotlight] == true ||
-        typeEnabledState[StepActionType.LineSpotlight]) {
+        typeEnabledState[StepActionType.LineSpotlight] == true) {
       c = addColor(c, yellow);
     }
     if (typeEnabledState[StepActionType.Show] == true ||
         typeEnabledState[StepActionType.Hide] == true) {
       c = addColor(c, blue);
     }
-    if (c != null) {
-      jss.set('div.cs-mark.${activeRegion.uniqClass}',
-          new JsObject.jsify({'background-color': c.toString()}));
-    }
+    jss.set('div.cs-mark.${region.uniqClass}',
+        new JsObject.jsify({'background-color': c.toString()}));
   }
 }
